use element selector when adding text assertions
