+++ extension/src/frontend/view/ElementInspector/ElementInspector.hooks.ts
@@ -1,91 +1,12 @@
-import { last } from 'lodash-es'
-import { useCallback, useEffect, useMemo, useState } from 'react'
+import { useCallback, useEffect, useState } from 'react'
 
-import { ElementSelector } from '@/schemas/recording'
-import { uuid } from '@/utils/uuid'
-import { ElementRole, getElementRoles } from 'extension/src/utils/aria'
-
-import { generateSelectors } from '../../../selectors'
 import { useGlobalClass } from '../GlobalStyles'
 import { useHighlightDebounce } from '../hooks/useHighlightDebounce'
 import { usePreventClick } from '../hooks/usePreventClick'
 import { Bounds, Position } from '../types'
-import { getElementBounds } from '../utils'
 
-function* getAncestors(element: Element) {
-  let current: Element | null = element.parentElement
-
-  while (current !== null) {
-    yield current
-
-    if (current === document.documentElement) {
-      break
-    }
-
-    current = current.parentElement
-  }
-}
-
-function findLabelFor(label: HTMLLabelElement): Element | null {
-  const id = label.getAttribute('for')
-
-  if (id === null) {
-    return null
-  }
-
-  return document.getElementById(id)
-}
-
-function findInChildren(label: HTMLLabelElement): Element | null {
-  return label.querySelector('input, select, textarea')
-}
-
-function findLabelledBy(label: HTMLLabelElement): Element | null {
-  if (label.id === '') {
-    return null
-  }
-
-  return document.querySelector(`[aria-labelledby="${label.id}"]`)
-}
-
-export function findRelatedInput(element: Element): TrackedElement | null {
-  const label = [...getAncestors(element)].find(
-    (ancestor) => ancestor instanceof HTMLLabelElement
-  )
-
-  if (label === undefined) {
-    return null
-  }
-
-  const input =
-    findLabelFor(label) ?? findInChildren(label) ?? findLabelledBy(label)
-
-  if (input === null) {
-    return null
-  }
-
-  return toTrackedElement(input)
-}
-
-export interface TrackedElement {
-  id: string
-  roles: ElementRole[]
-  selector: ElementSelector
-  target: Element
-  bounds: Bounds
-}
-
-function toTrackedElement(element: Element): TrackedElement {
-  const roles = getElementRoles(element)
-
-  return {
-    id: uuid(),
-    roles: [...roles],
-    selector: generateSelectors(element),
-    target: element,
-    bounds: getElementBounds(element),
-  }
-}
+import { usePinnedElement } from './hooks'
+import { toTrackedElement, TrackedElement } from './utils'
 
 function isInsideBounds(
   position: Position,
@@ -118,8 +39,9 @@ export function useInspectedElement() {
    *    the selection and removed when the user contracts the selection. If the
    *    stack is empty, then no element is pinned.
    */
-  const [pinnedEl, setPinnedElement] = useState<TrackedElement[]>([])
   const [hoveredEl, setHoveredEl] = useState<TrackedElement | null>(null)
+
+  const { selected, pinned, pin, unpin, expand, contract } = usePinnedElement()
 
   useGlobalClass('inspecting')
 
@@ -155,14 +77,14 @@ export function useInspectedElement() {
     }
   }, [])
 
-  const unpin = useCallback(() => {
+  const reset = useCallback(() => {
     setMousePosition({
       top: 0,
       left: 0,
     })
 
-    setPinnedElement([])
-  }, [])
+    unpin()
+  }, [unpin])
 
   usePreventClick({
     callback: (ev) => {
@@ -170,8 +92,8 @@ export function useInspectedElement() {
         return
       }
 
-      if (pinnedEl.length > 0) {
-        unpin()
+      if (pinned !== null) {
+        reset()
 
         return
       }
@@ -191,53 +113,18 @@ export function useInspectedElement() {
       }
 
       setMousePosition(position)
-      setPinnedElement([hoveredEl])
+      pin(hoveredEl)
     },
-    dependencies: [pinnedEl, hoveredEl, unpin],
+    dependencies: [pinned, hoveredEl, unpin],
   })
-
-  const expand = useMemo(() => {
-    const [head] = pinnedEl
-
-    if (head === undefined) {
-      return undefined
-    }
-
-    const parent = head.target.parentElement
-
-    if (parent === null || parent === document.documentElement) {
-      return undefined
-    }
-
-    return () => {
-      setPinnedElement((pinned) => {
-        return [toTrackedElement(parent), ...pinned]
-      })
-    }
-  }, [pinnedEl])
-
-  const contract = useMemo(() => {
-    const [head, ...tail] = pinnedEl
-
-    // If head is undefined, that means no element is pinned. If tail is
-    // empty, that means we're back at the intial element. In either case
-    // we can't decrease the selection any further.
-    if (head === undefined || tail.length === 0) {
-      return undefined
-    }
-
-    return () => {
-      setPinnedElement(tail)
-    }
-  }, [pinnedEl])
 
   const highlightedEl = useHighlightDebounce(hoveredEl)
 
   return {
-    pinned: last(pinnedEl) ?? null,
-    element: pinnedEl[0] ?? highlightedEl,
+    pinned,
+    element: selected ?? highlightedEl,
     mousePosition,
-    unpin,
+    reset,
     expand,
     contract,
   }
