+++ extension/src/frontend/view/TextSelectionPopover.tsx
@@ -1,5 +1,5 @@
 import { css } from '@emotion/react'
-import { useEffect, useState } from 'react'
+import { useState } from 'react'
 
 import { uuid } from '@/utils/uuid'
 
@@ -8,6 +8,7 @@ import { client } from '../routing'
 import { ElementPopover } from './ElementInspector/ElementPopover'
 import { TextAssertionEditor } from './ElementInspector/assertions/TextAssertionEditor'
 import { TextAssertionData } from './ElementInspector/assertions/types'
+import { useElementHighlight, usePinnedElement } from './ElementInspector/hooks'
 import { useGlobalClass } from './GlobalStyles'
 import { Overlay } from './Overlay'
 import { useTextSelection } from './TextSelectionPopover.hooks'
@@ -26,27 +27,28 @@ function TextSelectionPopoverContent({
   onAdd,
   onClose,
 }: TextSelectionPopoverContentProps) {
+  const { selected, expand, contract } = usePinnedElement(selection.element)
+
   const [assertion, setAssertion] = useState<TextAssertionData>({
     type: 'text',
-    selector: selection.selector.css,
+    selector: selection.element.selector.css,
     text: selection.text,
   })
 
-  useEffect(() => {
-    return () => {
-      client.send({
-        type: 'highlight-elements',
-        selector: null,
-      })
-    }
-  }, [])
+  const targetElement = selected ?? selection.element
+
+  useElementHighlight(targetElement)
 
   const handleChange = (assertion: TextAssertionData) => {
     setAssertion(assertion)
   }
 
   const handleSubmit = (assertion: TextAssertionData) => {
-    onAdd(assertion)
+    onAdd({
+      ...assertion,
+      selector: targetElement.selector.css,
+    })
+
     onClose()
   }
 
@@ -54,11 +56,16 @@ function TextSelectionPopoverContent({
     <ElementPopover
       open
       anchor={<Overlay bounds={selection.bounds} />}
-      header="Add text assertion"
+      header={
+        <ElementPopover.Selector
+          element={targetElement}
+          onExpand={expand}
+          onContract={contract}
+        />
+      }
       onOpenChange={onClose}
     >
       <TextAssertionEditor
-        canEditSelector
         assertion={assertion}
         onCancel={onClose}
         onChange={handleChange}
